Implement lazy mine placement
