+++ src/minesweeper/board/boardSetup.ts
@@ -1,8 +1,7 @@
-import { BOARD_SIZE, MINE_AMOUNT, DIRECTIONS } from "../constants";
-import type { Cell, Board } from "../types";
-import { isOutOfBounds } from "./utils";
+import { BOARD_SIZE } from "../constants";
+import type { Cell } from "../types";
 
-function createEmptyBoard() {
+export function createEmptyBoard() {
   return Array.from(Array(BOARD_SIZE.ROWS), () =>
     Array.from(
       Array(BOARD_SIZE.COLS),
@@ -13,33 +12,4 @@ function createEmptyBoard() {
       }),
     ),
   );
-}
-
-export function setup(): Board {
-  const board = createEmptyBoard();
-
-  let placed = 0;
-  while (placed < MINE_AMOUNT) {
-    const row = Math.floor(Math.random() * BOARD_SIZE.ROWS);
-    const col = Math.floor(Math.random() * BOARD_SIZE.COLS);
-
-    const cell = board[row][col];
-
-    if (cell.content === "mine") {
-      continue;
-    }
-
-    cell.content = "mine";
-    placed++;
-
-    for (const [rowOffset, colOffset] of DIRECTIONS) {
-      const newRow = row + rowOffset;
-      const newCol = col + colOffset;
-      if (!isOutOfBounds(newRow, newCol)) {
-        board[newRow][newCol].bombsAround++;
-      }
-    }
-  }
-
-  return board;
 }
