+++ src/minesweeper/board/engine.ts
@@ -1,5 +1,5 @@
 import type { Board } from "../types";
-import { DIRECTIONS } from "../constants";
+import { BOARD_SIZE, DIRECTIONS, MINE_AMOUNT } from "../constants";
 import { isOutOfBounds } from "./utils";
 
 export function toggleFlag(board: Board, row: number, col: number): Board {
@@ -33,6 +33,41 @@ export function revealCells(board: Board, row: number, col: number): Board {
 
   for (const [rowOffset, colOffset] of DIRECTIONS) {
     board = revealCells(board, row + rowOffset, col + colOffset);
+  }
+
+  return board;
+}
+
+export function placeMines(
+  board: Board,
+  safeRow: number,
+  safeCol: number,
+): Board {
+  let placed = 0;
+  while (placed < MINE_AMOUNT) {
+    const row = Math.floor(Math.random() * BOARD_SIZE.ROWS);
+    const col = Math.floor(Math.random() * BOARD_SIZE.COLS);
+
+    if (row === safeRow && col === safeCol) {
+      continue;
+    }
+
+    const cell = board[row][col];
+
+    if (cell.content === "mine") {
+      continue;
+    }
+
+    cell.content = "mine";
+    placed++;
+
+    for (const [rowOffset, colOffset] of DIRECTIONS) {
+      const newRow = row + rowOffset;
+      const newCol = col + colOffset;
+      if (!isOutOfBounds(newRow, newCol)) {
+        board[newRow][newCol].bombsAround++;
+      }
+    }
   }
 
   return board;
